requirements.txt
langgraph==0.6.2
langchain==0.3.27
langchain-community==0.3.27
langchain_openai-0.3.28
openai==1.98.0
langchain_core==2.2.3
pdfplumber==0.11.7
trafilatura==2.0.0
PyPDF2==3.0.1
pymupdf==1.26.3